Simplify experimentation files setup further
