--- app/mobile/features/experimentation/exposureLog.ts
@@ -1,24 +1,0 @@
-import { Experiment, Result } from "@growthbook/growthbook";
-import * as Sentry from "@sentry/react-native";
-
-import { reportExposure } from "@/service/experiments";
-
-export function recordExposure(
-  experiment: Experiment<unknown>,
-  result: Result<unknown>,
-): void {
-  void reportExposure({
-    experimentKey: experiment.key,
-    experimentName: experiment.name ?? "",
-    variationId: result.variationId,
-    variationKey: result.key,
-    variationName: result.name ?? "",
-    hashAttribute: result.hashAttribute,
-    hashValue: result.hashValue,
-    featureId: result.featureId ?? "",
-    inExperiment: result.inExperiment,
-    bucket: result.bucket,
-    hashUsed: result.hashUsed,
-    stickyBucketUsed: result.stickyBucketUsed,
-  }).catch((e) => Sentry.captureException(e));
-}
